refactor(application controller): refactor the application controller
